+++ helpers/RunLoadTest.ps1
@@ -17,8 +17,8 @@ Import-Module -Name $manifestPath -Force -PassThru
 
 $counter = 0
 while ($counter++ -lt $Count) {
-    $connection = Connect-TibberWebSocket -Verbose:$Verbose -Debug:$Debug
-    $subscription = Register-TibberLiveMeasurementSubscription -Connection $connection -HomeId '96a14971-525a-4420-aae9-e5aedaa129ff' -Verbose:$Verbose -Debug:$Debug
+    $connection = Connect-TibberWebSocket -HomeId '96a14971-525a-4420-aae9-e5aedaa129ff' -Verbose:$Verbose -Debug:$Debug
+    $subscription = Register-TibberLiveMeasurementSubscription -Connection $connection -Verbose:$Verbose -Debug:$Debug
     $result = Read-TibberWebSocket -Connection $connection -Callback {} -PackageCount 1 -Verbose:$Verbose -Debug:$Debug
     Write-Host "[$counter] Read $($result.NumberOfPackages) packages in $($result.ElapsedTimeInSeconds) seconds | Connection attempts: $($connection.ConnectionAttempts)"
     Unregister-TibberLiveMeasurementSubscription -Connection $connection -Subscription $subscription -Verbose:$Verbose -Debug:$Debug